Fixed !honor global storage duplication
There was an issue where !honor was storing multiple repeated entries in the global user. I fixed this for the Degens and EtherscanTransactions by the following lines:
```
        let Glob_Degens = new Set(data_global.Item.FellowDegens)
        let tmp_Glob_Degens = data_global.Item.FellowDegens
```
the Glob_Degens set can be searched in a for loop to the length of the Total_Degens variable which holds the degens that were involved in this honor. Using the Set.has function this becomes super efficient and we just append any new values to the tmp_Glob_Degens Array which has the last checked values for the global user. Then after the for loop we just rename the tmp_Glob_Degens to Global_Degens for consistency (not needed) and repeat this same process for the Transactions. (Lines 517-541 were added). 

Then in Lines 569 and 572 I simply call the Global_Degens and Global_tx in the global_params to save to the DynamoDB since we already know they only contain unique values of their contents because of the set.has loop